Intento de agregar un metodo para sumarle las ganancias pero ha sido intento fallido

+++ Proyectos/MotoGPPerezMarquezGabriel/src/main/java/com/salesianostriana/dam/motogpperezmarquezgabriel/model/Moto.java
@@ -21,5 +21,5 @@ public class Moto {
 	@Column(name = "tipo_motor")
 	private String tipoMotor;
 	@Column(name = "porcent_desgaste")
-	private int porcentDesgaste;
+	private double porcentDesgaste;
 }
